V5.3-004

+++ sr_port/dollarx.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- *	Copyright 2001, 2006 Fidelity Information Services, Inc	*
+ *	Copyright 2001, 2009 Fidelity Information Services, Inc	*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -206,7 +206,7 @@ void dollarx(io_desc *io_ptr, unsigned char *str, unsigned char *strtop)
 		io_ptr->dollar.x += total;
 #endif
 	} else
-		io_ptr->dollar.x += (strtop - str);
+		io_ptr->dollar.x += (unsigned int)(strtop - str);
 	if (io_ptr->dollar.x > io_ptr->width && io_ptr->wrap)
 	{
 		io_ptr->dollar.y += (io_ptr->dollar.x / io_ptr->width);
